myrequirements.txt
auto-sklearn==0.15.0
autopep8==2.0.1
everywhereml==0.2.17
influxdb==5.3.1
ipykernel==6.0.0
ipython==8.6.0
jupyter==1.0.0
matplotlib==3.6.2
mlflow[extras]==2.1.1
numpy==1.23.4
paho-mqtt==1.6.1
pandas==1.5.3
scikit-learn==0.24.2
scipy==1.10.1
seaborn==0.12.2
